+++ src/android/com/voxeet/toolkit/notification/CordovaIncomingNotification.java
@@ -11,33 +11,13 @@ import com.voxeet.uxkit.incoming.IncomingNotification;
 
 public class CordovaIncomingNotification extends IncomingNotification {
 
-    @NonNull
-    private final IncomingFullScreen incomingFullScreen;
-
     public CordovaIncomingNotification() {
         super();
-        incomingFullScreen = new IncomingFullScreen(CordovaIncomingCallActivity.class);
     }
 
     @Override
     public String getIncomingAcceptedClass(@NonNull Context context) {
         //not returning super since this activity will be responsible for that
         return CordovaIncomingFromNotificationActivity.class.getName(); //canonical would have returned name
-    }
-
-    @Override
-    public void onInvitation(@NonNull Context context, @NonNull InvitationBundle invitationBundle) {
-        CordovaRootViewProvider provider = VoxeetApplication.ROOT_VIEW_PROVIDER;
-        if (null != provider && null != provider.getCurrentActivity()) {
-            incomingFullScreen.onInvitation(context, invitationBundle);
-        } else super.onInvitation(context, invitationBundle);
-    }
-
-    @Override
-    public void onInvitationCanceled(@NonNull Context context, @NonNull String conferenceId) {
-        CordovaRootViewProvider provider = VoxeetApplication.ROOT_VIEW_PROVIDER;
-        if (null != provider && null != provider.getCurrentActivity()) {
-            incomingFullScreen.onInvitationCanceled(context, conferenceId);
-        } else super.onInvitationCanceled(context, conferenceId);
     }
 }