+++ src/ios/CDVVoxeet.m
@@ -113,26 +113,26 @@
 - (void)create:(CDVInvokedUrlCommand *)command {
     NSDictionary<NSString *,id> *options = [command.arguments objectAtIndex:0];
     
-    NSMutableDictionary *nativeOptions = [[NSMutableDictionary alloc] init];
-    [nativeOptions setValue:[options valueForKey:@"alias"] forKey:@"conferenceAlias"];
-    
+    // Create conference options.
+    VTConferenceOptions *conferenceOptions = [[VTConferenceOptions alloc] init];
+    conferenceOptions.alias = [options valueForKey:@"alias"];
+    conferenceOptions.pinCode = [options valueForKey:@"pinCode"];
     NSDictionary *params = [options valueForKey:@"params"];
     if (params) {
-        NSMutableDictionary *nativeOptionsParams = [[NSMutableDictionary alloc] init];
-        [nativeOptionsParams setValue:[params valueForKey:@"ttl"] forKey:@"ttl"];
-        [nativeOptionsParams setValue:[params valueForKey:@"rtcpMode"] forKey:@"rtcpMode"];
-        [nativeOptionsParams setValue:[params valueForKey:@"mode"] forKey:@"mode"];
-        [nativeOptionsParams setValue:[params valueForKey:@"videoCodec"] forKey:@"videoCodec"];
-        [nativeOptions setValue:nativeOptionsParams forKey:@"params"];
-        
-        if ([params valueForKey:@"liveRecording"]) {
-            [nativeOptions setValue:@{@"liveRecording": [params valueForKey:@"liveRecording"]} forKey:@"metadata"];
-        }
+        conferenceOptions.params.liveRecording = [params valueForKey:@"liveRecording"];
+        conferenceOptions.params.rtcpMode = [params valueForKey:@"rtcpMode"];
+        conferenceOptions.params.stats = [params valueForKey:@"stats"];
+        conferenceOptions.params.ttl = [params valueForKey:@"ttl"];
+        conferenceOptions.params.videoCodec = [params valueForKey:@"videoCodec"];
+        conferenceOptions.params.dolbyVoice = [params valueForKey:@"dolbyVoice"];
     }
     
     dispatch_async(dispatch_get_main_queue(), ^{
-        [VoxeetSDK.shared.conference createWithParameters:nativeOptions success:^(NSDictionary<NSString *,id> *response) {
-            [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:response] callbackId:command.callbackId];
+        [VoxeetSDK.shared.conference createWithOptions:conferenceOptions success:^(VTConference *conference) {
+            NSDictionary *result = @{@"conferenceId": conference.id,
+                                     @"conferenceAlias": conference.alias,
+                                     @"isNew": [NSNumber numberWithBool:conference.isNew]};
+            [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result] callbackId:command.callbackId];
         } fail:^(NSError *error) {
             [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];
         }];
@@ -141,25 +141,41 @@
 
 - (void)join:(CDVInvokedUrlCommand *)command {
     NSString *conferenceID = [command.arguments objectAtIndex:0];
-    NSDictionary<NSString *,id> *options = nil;
+    BOOL isListener = NO;
+    BOOL defaultVideo = VoxeetSDK.shared.conference.defaultVideo; /* Monkey patch with listener mode */
     if ([command.arguments count] > 1) {
-        options = [command.arguments objectAtIndex:1];
-    }
-    
-    NSMutableDictionary *nativeOptions = [[NSMutableDictionary alloc] init];
-    [nativeOptions setValue:[options valueForKey:@"alias"] forKey:@"conferenceAlias"];
-    
-    NSDictionary *user = [options valueForKey:@"user"];
-    if (user) {
-        [nativeOptions setValue:[user valueForKey:@"type"] forKey:@"participantType"];
+        NSDictionary<NSString *, id> *options = [command.arguments objectAtIndex:1];
+        NSDictionary *user = [options valueForKey:@"user"];
+        if (user != nil) {
+            NSString *type = [user valueForKey:@"type"];
+            if (type != nil && [type isEqual:@"listener"]) {
+                isListener = YES;
+            }
+        }
     }
     
     dispatch_async(dispatch_get_main_queue(), ^{
-        BOOL video = VoxeetSDK.shared.conference.defaultVideo;
-        [VoxeetSDK.shared.conference joinWithConferenceID:conferenceID video:video userInfo:nativeOptions success:^(NSDictionary<NSString *,id> *response) {
-            [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:response] callbackId:command.callbackId];
-        } fail:^(NSError *error) {
-            [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];
+        [VoxeetSDK.shared.conference fetchWithConferenceID:conferenceID completion:^(VTConference *conference) {
+            if (!isListener) {
+                VTJoinOptions *options = [[VTJoinOptions alloc] init];
+                options.constraints.video = VoxeetSDK.shared.conference.defaultVideo;
+                [VoxeetSDK.shared.conference joinWithConference:conference options:options success:^(VTConference *conference2) {
+                    NSDictionary *result = @{@"conferenceId": conference2.id, @"conferenceAlias": conference2.alias};
+                    [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result] callbackId:command.callbackId];
+                } fail:^(NSError *error) {
+                    [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];
+                }];
+            } else {
+                VoxeetSDK.shared.conference.defaultVideo = NO;
+                [VoxeetSDK.shared.conference listenWithConference:conference success:^(VTConference *conference2) {
+                    VoxeetSDK.shared.conference.defaultVideo = defaultVideo;
+                    NSDictionary *result = @{@"conferenceId": conference2.id, @"conferenceAlias": conference2.alias};
+                    [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsDictionary:result] callbackId:command.callbackId];
+                } fail:^(NSError *error) {
+                    VoxeetSDK.shared.conference.defaultVideo = defaultVideo;
+                    [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];
+                }];
+            }
         }];
     });
 }
@@ -394,26 +410,33 @@
 - (void)startConference:(CDVInvokedUrlCommand *)command { /* Deprecated */
     NSString *confAlias = [command.arguments objectAtIndex:0];
     NSArray *participants = [command.arguments objectAtIndex:1];
-    NSMutableArray *userIDs = [[NSMutableArray alloc] init];
+    NSMutableArray<VTParticipantInfo *> *participantInfos = [[NSMutableArray alloc] init];
     
     for (NSDictionary *participant in participants) {
-        [userIDs addObject:[participant objectForKey:@"externalId"]];
+        NSString *externalID = [participant objectForKey:@"externalId"];
+        NSString *name = [participant objectForKey:@"name"];
+        NSString *avatarURL = [participant objectForKey:@"avatarUrl"];
+        
+        VTParticipantInfo *participantInfo = [[VTParticipantInfo alloc] initWithExternalID:externalID name:name avatarURL:avatarURL];
+        [participantInfos addObject:participantInfo];
     }
     
+    // Create conference options.
+    VTConferenceOptions *conferenceOptions = [[VTConferenceOptions alloc] init];
+    conferenceOptions.alias = confAlias;
+    
     dispatch_async(dispatch_get_main_queue(), ^{
-        [VoxeetSDK.shared.conference createWithParameters:@{@"conferenceAlias": confAlias} success:^(NSDictionary<NSString *,id> *response) {
-            NSString *confID = response[@"conferenceId"];
-            BOOL isNew = response[@"isNew"];
-            BOOL video = VoxeetSDK.shared.conference.defaultVideo;
-            
-            [VoxeetSDK.shared.conference joinWithConferenceID:confID video:video userInfo:nil success:^(NSDictionary<NSString *,id> *response) {
+        [VoxeetSDK.shared.conference createWithOptions:conferenceOptions success:^(VTConference *conference) {
+            VTJoinOptions *joinOptions = [[VTJoinOptions alloc] init];
+            joinOptions.constraints.video = VoxeetSDK.shared.conference.defaultVideo;
+            [VoxeetSDK.shared.conference joinWithConference:conference options:joinOptions success:^(VTConference *conference2) {
                 [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_OK] callbackId:command.callbackId];
             } fail:^(NSError *error) {
                 [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];
             }];
             
-            if (isNew) {
-                [VoxeetSDK.shared.conference inviteWithConferenceID:confID externalIDs:userIDs completion:^(NSError *error) {}];
+            if (conference.isNew) {
+                [VoxeetSDK.shared.notification inviteWithConference:conference participantInfos:participantInfos completion:nil];
             }
         } fail:^(NSError *error) {
             [self.commandDelegate sendPluginResult:[CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageAsString:error.description] callbackId:command.callbackId];